requirements.txt
altgraph==0.17
future==0.18.2
pefile==2019.4.18
pyinstaller==4.2
pyinstaller-hooks-contrib==2021.1
pypiwin32==223
pywin32==300
pywin32-ctypes==0.2.0
six==1.15.0
websocket-client==0.58.0
win10toast==0.9
